move data handlers, rename dataconfig->dataoptions
