+++ genesis/sensors/__init__.py
@@ -1,4 +1,4 @@
 from .base_sensor import Sensor
 from .camera import Camera
 from .tactile import RigidContactSensor, RigidContactForceSensor, RigidContactForceGridSensor
-from .data_collector import DataCollector, DataOutType
+from .data_collector import DataCollector, DataRecordingOptions
